chore: more action directory changes
